Resolve "Optimeze the cache of the pipeline jobs"
Currently, the pipeline's cache is created for each branch. However, this appears redundant if nothing changes in the node_modules directory.

An improvement would be to revalidate the cache only when changes occur in the `yarn.lock` file.

Proposed Solution: Use the `yarn.lock` file as the key for the cache.